MANPAGER2: Avoid conflict with the stock MANPAGER
Vim comes with a MANPAGER plugin in $VIMRUNTIME/plugin/manpager.vim.
$VIMRUNTIME plugin defines a MANPAGER command already.  Redefining the
same command could be error-prone.  And as of a recent update

    commit d592deb336523a5448779ee3d4bba80334cff1f7
    Author: Bram Moolenaar <Bram@vim.org>
    Date:   Fri Jun 17 15:42:40 2022 +0100

        Update runtime files

MANPAGER plugin in $VIMRUNTIME defines MANPAGER command without a
consideration of whether it already exists or not.  Effectively
generating an error on every startup if `vim-manpager` plugin is also
loaded.